use a shared socket hook
